+++ hdhomerun_video.h
@@ -54,7 +54,7 @@ extern LIBHDHOMERUN_API void hdhomerun_video_destroy(struct hdhomerun_video_sock
 /*
  * Configure to send a keepalive packet every second.
  */
-extern LIBHDHOMERUN_API void hdhomerun_video_set_keepalive(struct hdhomerun_video_sock_t *vs, uint32_t remote_addr, uint32_t remote_port, uint32_t lockkey);
+extern LIBHDHOMERUN_API void hdhomerun_video_set_keepalive(struct hdhomerun_video_sock_t *vs, uint32_t remote_addr, uint16_t remote_port, uint32_t lockkey);
 
 /*
  * Get the port the socket is listening on.
